rename package to iriscasttools
