+++ setup.py
@@ -1,12 +1,12 @@
 from setuptools import setup, find_packages
 
-VERSION = '0.0.2'
-DESCRIPTION = 'python package for querying IPMI tool'
+VERSION = '0.0.1'
+DESCRIPTION = 'python package for IRISCAST tools'
 
-LONG_DESCRIPTION = """Python package to query IPMITool to get power statistics"""
+LONG_DESCRIPTION = """Python package to get power and CPU/RAM usage statistics"""
 
 setup (
-        name="iriscastipmi",
+        name="iriscasttools",
         version=VERSION,
         author="Anish Mudaraddi",
         author_email="<anish.mudaraddi@stfc.ac.uk>",
@@ -14,7 +14,6 @@ setup (
         long_description=LONG_DESCRIPTION,
         packages=find_packages(),
         python_requires='>=3',
-        install_requires=[], # add any additional packages that
-        # needs to be installed along with your package. Eg: 'caer'
+        install_requires=[],
         keywords=['python']
 )
